+++ wayland-client/CHANGELOG.md
@@ -18,6 +18,7 @@
 
 #### Additions
 - Add `GlobalList::bind_specific`
+- Add `GlobalList::bind_all`
 - Updated Wayland core protocol to 1.26
 - Implement `AsRawFd` for `Connection` and `EventQueue`